pyproject.toml
[tool.poetry]
name = "django-ide-app"
version = "0.1.0"
description = ""
authors = ["Daniel0Marsh <danielmarsh1000@protonmail.com>"]
readme = "README.md"

packages = [
    { include = "source" }
]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
docker = "^7.0.0"
pillow = "^11.0.0"
markdown = "^3.7"
social-auth-app-django = "^5.4.2"
aiohttp = "^3.11.11"
pygithub = "^2.5.0"
environ = "^1.0"
django-environ = "^0.12.0"
gitpython = "^3.1.44"
python-decouple = "^3.8"
stripe = "^11.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
